introduce the `(print)` procedure (#489)
* rename #%top-level-print to #%top-level-display and #%print to #%display

if i want to add a (print) function, it is quite unfortunate, that the
namespace is already filled with functions, that are called something
with print, that are then only used by (display). so rename them.

* introduce the (print) procedure

* add some display/print tests

most of these are still commented out, while the display impl is not
where i want it to be, but they can be here regardless.

* add a println and use it in the repl