misc port improvements (#505)
* actually check for WouldBlock in read_byte

* remove unused bool return from read_bytes_amt

i wanted to add some (unrelated to this pr) features, and to do that i
had to refactor the read_bytes_amt fn, where i stumbled upon the bool
return parameter, which i didn't know what it was supposed to indicate.
so i checked where it was used to figure it out, und it turns out it's
not used at all, so i can just remove it.

* add some port tests for (read-line)

i forgot to add tests for it when i added it, so here they are, a little
delayed.

* improve the read_bytes_amt fn

the `read` function on a Read is allowed to just simply terminate before
filling the buf, even if it *could* fill the buf, which would be
undesirable.

additionally, a call to write should not return an error of kind
`UnexpectedEof`, it should return an Ok(0) and it is the duty of the
caller to then interpret that as an eof, in which case `read_exact` will
return that kind of error.

the current implementation is a slightly modified impl of the default
`read_exact` fn, as i sadly cannot call that one, as that does not give
me the amount of bytes read, in addition to the state of the input `buf`
not being specified if the function returns an error or reaches eof.